add SEO metadata to all major pages (landing, quiz, dashboard, chart/new, compatibility, login, moon, layout OG)
