+++ src/app/(main)/moon/page.tsx
@@ -212,4 +212,8 @@ function CalendarSkeleton() {
 export const metadata = {
   title: 'Місячний Календар | Зоря',
   description: 'Фази Місяця, періоди «Місяць без курсу» та персональні місячні транзити. Дізнайтеся коли краще починати справи і коли варто почекати.',
+  openGraph: {
+    title: 'Місячний Календар | Зоря',
+    description: 'Фази Місяця, періоди Місяць без курсу та персональні місячні транзити.',
+  },
 };
